# src/main/java/com/lacuisine/recipe/entity/Ingredients.java
package com.lacuisine.recipe.entity;

import java.io.Serializable;
import javax.persistence.CascadeType;
import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.FetchType;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.JoinColumn;
import javax.persistence.ManyToOne;
import javax.persistence.Table;
import org.apache.commons.lang3.builder.ToStringExclude;
import lombok.Getter;
import lombok.Setter;



@Table(name = "ingredients")
@Setter
@Getter
@Entity
public class Ingredients implements Serializable {

  private static final long serialVersionUID = 1L;

  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  private Long id;

  @Column(name = "ingredient")
  private String ingredient;

  @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
  @JoinColumn(name = "recipeId")
  @ToStringExclude
  private RecipeEntity recipeEntity;

}